moved pure functions to McRefCommon.cpp
