+++ src/CladeStats.cpp
@@ -42,7 +42,7 @@ void computeCladeNumCoals(){
 void computeCladeNumCoals_rec(int pop){
 	int leftSon, rightSon;
 
-	if (isLeafPopulation(pop)){
+	if (isLeaf(pop)){
 		clade_stats[pop].num_coals_total = genetree_stats_total.num_coals[pop];
 	} else{
 		leftSon = dataSetup.popTree->pops[pop]->sons[LEFT]->id;
@@ -66,7 +66,7 @@ void computeCladeCoalStats(){
 void computeCladeCoalStats_rec(int clade, int gen) {
 	int leftSon, rightSon;
 
-	if (isLeafPopulation(clade)){
+	if (isLeaf(clade)){
 		fillupLeafCladeStats(clade, gen);
 	} else{
 		leftSon = dataSetup.popTree->pops[clade]->sons[LEFT]->id;
@@ -238,20 +238,3 @@ void initSpecificCladeStats(int clade){
 void freeCladeMem(){
 }
 
-
-#define TRUE 1 // TODO - move this to common place with comb stats
-#define FALSE 0
-int	isLeafPopulation(int pop){ // TODO - move this to common place with comb stats
-	Population *population, *left_son, *right_son;
-
-	population = dataSetup.popTree->pops[pop];
-
-	left_son = population->sons[LEFT];
-	right_son = population->sons[RIGHT];
-
-	if (left_son || right_son){
-		return FALSE;
-	} else {
-		return TRUE;
-	}
-}
